Refactoring fetchFileMetadata
