+++ src/test/integration/completeFlow.test.ts
@@ -243,20 +243,20 @@ describe('Should run a complete node flow.', () => {
   //   )
   // })
 
-  it('should get file info with did', async () => {
-    const fileInfoRequest: FileInfoRequest = {
-      did: assetDID,
-      serviceId
-    }
-    const fileInfo = await storage.getFileInfo(fileInfoRequest, p2pNode)
-    console.log('fileInfo', fileInfo)
+  // it('should get file info with did', async () => {
+  //   const fileInfoRequest: FileInfoRequest = {
+  //     did: assetDID,
+  //     serviceId
+  //   }
+  //   const fileInfo = await storage.getFileInfo(fileInfoRequest)
+  //   console.log('fileInfo', fileInfo)
 
-    assert(fileInfo[0].valid, 'File info is valid')
-    expect(fileInfo[0].contentLength).to.equal('417')
-    expect(fileInfo[0].contentType).to.equal('text/plain; charset=utf-8')
-    expect(fileInfo[0].name).to.equal('algo.js')
-    expect(fileInfo[0].type).to.equal('url')
-  })
+  //   assert(fileInfo[0].valid, 'File info is valid')
+  //   expect(fileInfo[0].contentLength).to.equal('417')
+  //   expect(fileInfo[0].contentType).to.equal('text/plain; charset=utf-8')
+  //   expect(fileInfo[0].name).to.equal('algo.js')
+  //   expect(fileInfo[0].type).to.equal('url')
+  // })
 
   it('should start an order', async function () {
     this.timeout(15000) // Extend default Mocha test timeout
